app/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0.post1
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-genai>=0.3.0
structlog>=23.2.0
python-dotenv>=1.0.0
greenlet>=3.0.3
